Reintegrate LayersVisibilityManager into LayerService
