+++ src/app/modules/layer-solar-system/services/visibility/paths-visibility-manager.service.spec.ts
@@ -1,7 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 import { LayerService } from '#core/services/layer.service';
 import { ThemeService } from '#core/services/theme.service';
-import { LayersVisibilityManagerService } from '#core/services/visibility/layers-visibility-manager.service';
 import { mockedLayers } from '#core/test-utils/mocked-layers.spec';
 import { LayerSolarSystemModule } from '#layer-solar-system/layer-solar-system.module';
 import { SolarSystem } from '#layer-solar-system/model/solar-system';
@@ -13,7 +12,6 @@ describe('PathsVisibilityManagerService', () => {
 
   const solarSystem = SolarSystem.CODE;
   let manager: PathsVisibilityManagerService;
-  let layersManager: LayersVisibilityManagerService;
   let layerService: LayerService;
 
   beforeEach(() => {
@@ -21,14 +19,12 @@ describe('PathsVisibilityManagerService', () => {
       imports: [LayerSolarSystemModule],
       providers: [
         LayerService,
-        LayersVisibilityManagerService,
         PathsVisibilityManagerService,
         ThemeService
       ]
     });
     layerService = TestBed.inject(LayerService);
     manager = TestBed.inject(PathsVisibilityManagerService);
-    layersManager = TestBed.inject(LayersVisibilityManagerService);
   });
 
   const loadSolarSystemLayer = (): SolarSystem => {
@@ -36,36 +32,36 @@ describe('PathsVisibilityManagerService', () => {
     const provider = TestBed.inject(SolarSystemProvidersService);
     const layer = provider.getRenderableLayer(model);
     layerService.registerLayer(layer);
-    layersManager.setVisible(solarSystem, true);
+    layerService.setVisible(solarSystem, true);
     return layer;
   };
 
   describe('showPaths should', () => {
 
     it('have no effect if the layer does not exist', () => {
-      spyOn(layersManager, 'setVisible');
+      spyOn(layerService, 'setVisible');
 
       manager.setPathsVisible(true);
-      expect(layersManager.setVisible).toHaveBeenCalledTimes(0);
+      expect(layerService.setVisible).toHaveBeenCalledTimes(0);
     });
 
     it('if the arg is true, trigger the showing of the solar system layer and trajectories', () => {
       const layer = loadSolarSystemLayer();
-      spyOn(layersManager, 'setVisible');
+      spyOn(layerService, 'setVisible');
       spyOn(layer, 'setTrajectoriesVisible');
 
       manager.setPathsVisible(true);
-      expect(layersManager.setVisible).toHaveBeenCalledTimes(2);
+      expect(layerService.setVisible).toHaveBeenCalledTimes(2);
       expect(layer.setTrajectoriesVisible).toHaveBeenCalledTimes(1);
     });
 
     it('if the arg is false, trigger the hiding of the solar system layer and trajectories', () => {
       const layer = loadSolarSystemLayer();
-      spyOn(layersManager, 'setVisible');
+      spyOn(layerService, 'setVisible');
       spyOn(layer, 'setTrajectoriesVisible');
 
       manager.setPathsVisible(false);
-      expect(layersManager.setVisible).toHaveBeenCalledTimes(2);
+      expect(layerService.setVisible).toHaveBeenCalledTimes(2);
       expect(layer.setTrajectoriesVisible).toHaveBeenCalledTimes(1);
     });
